python-starter/requirements.txt
anytree==2.8.0
app-common-python==0.2.1
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
cachetools==5.2.0
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
crc-bonfire==4.4.0
Django==4.0.5
django-environ==0.9.0
flake8==4.0.1
future==0.18.2
google-auth==2.8.0
gql==3.0.0a6
graphql-core==3.1.7
idna==3.3
iniconfig==1.1.1
junitparser==2.6.0
kafka-python==2.0.2
kubernetes==24.2.0
mccabe==0.6.1
minio==7.1.9
multidict==6.0.2
oauthlib==3.2.0
ocviapy==1.1.2
packaging==21.3
parsedatetime==2.6
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
requests==2.28.0
requests-oauthlib==1.3.1
rsa==4.8
sh==1.14.2
six==1.16.0
sqlparse==0.4.2
tabulate==0.8.10
tomli==2.0.1
urllib3==1.26.9
wait-for==1.2.0
websocket-client==1.3.3
yapf==0.32.0
yarl==1.7.2
